feat: move useAppTheme to design-system
