+++ packages/design-system/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useAppTheme";
